+++ fuse/eosfs.c
@@ -407,7 +407,7 @@ eosdfs_chown (const char* path, uid_t uid, gid_t gid)
 static int
 eosdfs_truncate (const char* path, off_t size)
 {
-  fprintf(stderr, "[%s] path=%s, size=%lli\n", __FUNCTION__, path, size);
+  fprintf(stderr, "[%s] path=%s, size=%lli\n", __FUNCTION__, path, (long long int)size);
   char rootpath[4096];
   unsigned long rinode = 0;
   eosatime = time (0);
@@ -501,7 +501,7 @@ eosdfs_read (const char* path,
              struct fuse_file_info* fi)
 {
   fprintf (stderr, "[%s] path=%s, offset=%llii, length=%li\n", 
-           __FUNCTION__, path, offset, size);
+           __FUNCTION__, path, (long long int)offset, size);
   eosatime = time (0);
   struct fd_user_info* info = (fd_user_info*) fi->fh;
   int res = xrd_pread (info->fd, buf, size, offset);
@@ -529,7 +529,7 @@ eosdfs_write (const char* path,
               struct fuse_file_info* fi)
 {
   fprintf (stderr, "[%s] path=%s, offset=%lli, lenght=%li\n", 
-           __FUNCTION__, path, offset, size);
+           __FUNCTION__, path, (long long int)offset, size);
 
   // File already existed. FUSE uses eosdfs_open() and eosdfs_truncate()
   // to open and truncate a file before calling eosdfs_write()
@@ -615,7 +615,7 @@ eosdfs_ftruncate(const char* path,
                  off_t size, 
                  struct fuse_file_info* fi)
 {
-  fprintf (stderr, "[%s] path=%s, size=%lli\n", __FUNCTION__, path, size);
+  fprintf (stderr, "[%s] path=%s, size=%lli\n", __FUNCTION__, path, (long long int)size);
   struct fd_user_info* info = (fd_user_info*) fi->fh;
   int res = xrd_truncate (info->fd, size);
